# swagger.yaml
openapi: 3.0.0
info:
  title: SelectLine ERP API
  description: API for interacting with SelectLine ERP data
  version: 1.0.0
  contact:
    email: support@example.com
servers:
  - url: /api
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "An error occurred"
    ValidationError:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              value:
                type: string
              msg:
                type: string
              param:
                type: string
              location:
                type: string
...
    TransformationRequest:
      type: object
      required:
        - sourceFormat
        - targetFormat
        - data
      properties:
        sourceFormat:
          type: string
          enum: ["selectline", "salesforce"]
          example: "selectline"
        targetFormat:
          type: string
          enum: ["selectline", "salesforce"]
          example: "salesforce"
        data:
          type: object
          description: The data to transform
          example: {
            "nummer": "K10001",
            "name": "Musterfirma GmbH"
          }
    TransformationResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: The transformed data
          example: {
            "AccountNumber": "K10001",
            "Name": "Musterfirma GmbH"
          }
    WebhookRegistration:
      type: object
      required:
        - url
        - events
      properties:
        url:
          type: string
          example: "https://your-service.com/webhook"
        events:
          type: array
          items:
            type: string
            enum: ["customer.created", "customer.updated", "order.created", "order.updated"]
          example: ["customer.created", "order.created"]
        secret:
          type: string
          description: Optional secret for webhook signature validation
          example: "your-webhook-secret"
...
security:
  - bearerAuth: []
paths:
...
  /transform:
    post:
      tags:
        - Transformation
      summary: Transform data between formats
      description: Transform data between SelectLine and Salesforce formats
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransformationRequest'
      responses:
        '200':
          description: Data transformed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransformationResponse'
        '400':
          description: Invalid request format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /webhooks:
    post:
      tags:
        - Webhooks
      summary: Register a new webhook
      description: Register a URL to receive webhook events
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookRegistration'
      responses:
        '201':
          description: Webhook registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookRegistration'
        '400':
          description: Invalid webhook configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Webhooks
      summary: List registered webhooks
      description: Get a list of all registered webhooks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebhookRegistration'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...